fix: code review - correcoes criticas de bugs e qualidade
database.cpp:
- substituir goto por loop com limite (MAX_QUERY_RETRIES=10) em storeQuery()
- my_bool -> bool (API removida no MySQL >= 8.0)
- new char[]/delete[] -> std::vector<char> em escapeBlob() (memory leak)

connection.cpp:
- descomentar mutex em getIP() - race condition critica com closeSocket()

configmanager.cpp:
- lua_tonumber -> lua_tointeger para leitura de inteiros do Lua

iologindata.cpp:
- blessings load: uint16_t -> uint8_t (consistente com TINYINT UNSIGNED)
- blessings save: uint32_t -> uint16_t (evita char cast em operator<<)

otserv.cpp:
- reordenar serverIPs: IP do config.lua primeiro, Docker IP por ultimo
- adicionar TODOs documentando gethostbyname() deprecated (POSIX 2001)

protocollogin.cpp:
- guard serverIPs.empty() antes de acessar [0] (UB/crash na inicializacao)

docker-compose.yml:
- healthcheck com senha root (-pyour_docker_password)
- remover atributo version obsoleto

CMakeLists.txt / cotire.cmake:
- cmake_minimum_required 3.5 -> 3.10
- C++11 explicito, MSVC compat, remover FindCXX11 deprecated

+++ src/iologindata.cpp
@@ -257,7 +257,9 @@ bool IOLoginData::loadPlayer(Player* player, DBResult_ptr result)
 
 	player->soul = result->getNumber<uint16_t>("soul");
 	player->capacity = result->getNumber<uint32_t>("cap") * 100;
-	player->blessings = result->getNumber<uint16_t>("blessings");
+	// FIX 🟡: blessings is stored as TINYINT UNSIGNED in the DB (max 255).
+	// Loading as uint16_t was unnecessarily wide; uint8_t matches the column type.
+	player->blessings = result->getNumber<uint8_t>("blessings");
 
 	unsigned long conditionsSize;
 	const char* conditions = result->getStream("conditions", conditionsSize);
@@ -759,7 +761,9 @@ bool IOLoginData::savePlayer(Player* player)
 	if (!player->isOffline()) {
 		query << "`onlinetime` = `onlinetime` + " << (time(nullptr) - player->lastLoginSaved) << ',';
 	}
-	query << "`blessings` = " << static_cast<uint32_t>(player->blessings);
+	// FIX 🟡: Cast to uint8_t instead of uint32_t — consistent with the DB column
+	// type (TINYINT UNSIGNED) and the load side above. uint32_t was over-wide.
+	query << "`blessings` = " << static_cast<uint16_t>(player->blessings);
 	query << " WHERE `id` = " << player->getGUID();
 
 	DBTransaction transaction;
